task3/requirements.txt
# =============================================================================
# Step 1 — 5-min → hourly aggregation (aggregate_to_hourly.py)
# Requires NVIDIA GPU + CUDA. Best run on Google Colab GPU runtime.
# Install RAPIDS cuDF: https://rapids.ai/start.html
# =============================================================================
cudf-cu12>=24.0        # GPU-accelerated DataFrame (RAPIDS) — version depends on CUDA
pandas>=2.0.0
numpy>=1.24.0

# =============================================================================
# Step 2 — Feature engineering (Feature_engineering_ipynb.ipynb)
# Runs inside Snowflake Notebooks — no local installation needed.
# =============================================================================

# =============================================================================
# Step 3 — Model training (trening_modelnn.py / best_model_inference.py)
# Requires NVIDIA GPU (A100/H100 recommended). CPU also works but is slower.
# =============================================================================

# PyTorch — install with CUDA support:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0

# ML / data
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
optuna>=3.4.0

# Utilities
pathlib2>=2.3.7        # built-in in Python 3.4+, listed for clarity
